requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
pydantic>=2.0.0

# For Replicate deployment
cog>=0.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
ruff>=0.1.0